+++ lib/src/controls/navigation/navigation_view/view.dart
@@ -19,9 +19,10 @@ const double _kDefaultAppBarHeight = 50.0;
 /// for your app. It adapts to a variety of screen sizes and
 /// supports both top and left navigation styles.
 ///
-/// ![](https://docs.microsoft.com/en-us/windows/uwp/design/controls-and-patterns/images/nav-view-header.png)
+/// ![NavigationView Preview](https://docs.microsoft.com/en-us/windows/uwp/design/controls-and-patterns/images/nav-view-header.png)
 ///
 /// See also:
+///
 ///   * [NavigationBody], a widget that implement fluent
 ///     transitions into [NavigationView]
 ///   * [NavigationPane], the pane used by [NavigationView],
@@ -104,6 +105,18 @@ class NavigationViewState extends State<NavigationView> {
   bool _compactOverlayOpen = false;
 
   int oldIndex = 0;
+
+  PaneDisplayMode? _autoDisplayMode;
+
+  /// Gets the current display mode. If it's automatic, it'll adapt to the other
+  /// display modes according to the current available space.
+  PaneDisplayMode get displayMode {
+    if (widget.pane?.displayMode == PaneDisplayMode.auto) {
+      return _autoDisplayMode ?? PaneDisplayMode.minimal;
+    }
+
+    return widget.pane?.displayMode ?? PaneDisplayMode.minimal;
+  }
 
   @override
   void initState() {
@@ -201,16 +214,15 @@ class NavigationViewState extends State<NavigationView> {
         double width = consts.biggest.width;
         if (width.isInfinite) width = MediaQuery.of(context).size.width;
 
-        late PaneDisplayMode autoDisplayMode;
         if (width <= 640) {
-          autoDisplayMode = PaneDisplayMode.minimal;
+          _autoDisplayMode = PaneDisplayMode.minimal;
         } else if (width >= 1008) {
-          autoDisplayMode = PaneDisplayMode.open;
+          _autoDisplayMode = PaneDisplayMode.open;
         } else if (width > 640) {
-          autoDisplayMode = PaneDisplayMode.compact;
+          _autoDisplayMode = PaneDisplayMode.compact;
         }
 
-        displayMode = autoDisplayMode;
+        displayMode = _autoDisplayMode!;
       }
 
       assert(displayMode != PaneDisplayMode.auto);
@@ -251,15 +263,17 @@ class NavigationViewState extends State<NavigationView> {
         final pane = widget.pane!;
         if (pane.customPane != null) {
           paneResult = Builder(builder: (context) {
-            return pane.customPane!.build(
-              context,
-              NavigationPaneWidgetData(
-                appBar: appBar,
-                content: ClipRect(child: widget.content),
-                listKey: _listKey,
-                paneKey: _panelKey,
-                scrollController: scrollController,
-                pane: pane,
+            return PaneScrollConfiguration(
+              child: pane.customPane!.build(
+                context,
+                NavigationPaneWidgetData(
+                  appBar: appBar,
+                  content: ClipRect(child: widget.content),
+                  listKey: _listKey,
+                  paneKey: _panelKey,
+                  scrollController: scrollController,
+                  pane: pane,
+                ),
               ),
             );
           });
@@ -312,8 +326,7 @@ class NavigationViewState extends State<NavigationView> {
                 setState(() => _compactOverlayOpen = !_compactOverlayOpen);
               }
 
-              final openSize =
-                  pane.size?.openWidth ?? _kOpenNavigationPanelWidth;
+              final openSize = pane.size?.openWidth ?? kOpenNavigationPaneWidth;
 
               final bool openedWithoutOverlay =
                   _compactOverlayOpen && consts.maxWidth / 2.5 > openSize;
@@ -325,8 +338,7 @@ class NavigationViewState extends State<NavigationView> {
                   top: widget.appBar?.height ?? 0.0,
                   start: openedWithoutOverlay
                       ? openSize
-                      : pane.size?.compactWidth ??
-                          _kCompactNavigationPanelWidth,
+                      : pane.size?.compactWidth ?? kCompactNavigationPaneWidth,
                   end: 0,
                   bottom: 0,
                   child: content,
@@ -345,6 +357,7 @@ class NavigationViewState extends State<NavigationView> {
                       ),
                     ),
                   ),
+                // appBar,
                 PaneScrollConfiguration(
                   child: () {
                     if (openedWithoutOverlay) {
@@ -365,30 +378,27 @@ class NavigationViewState extends State<NavigationView> {
                         ),
                       );
                     } else if (_compactOverlayOpen) {
-                      return ColoredBox(
-                        color: Colors.black,
-                        child: Mica(
-                          key: _overlayKey,
-                          backgroundColor: _overlayBackgroundColor(),
-                          elevation: 10.0,
-                          child: Container(
-                            decoration: BoxDecoration(
-                              border: Border.all(
-                                color: const Color(0xFF6c6c6c),
-                                width: 0.15,
-                              ),
-                              borderRadius: BorderRadius.circular(8.0),
+                      return Mica(
+                        key: _overlayKey,
+                        backgroundColor: _overlayBackgroundColor(),
+                        elevation: 10.0,
+                        child: Container(
+                          decoration: BoxDecoration(
+                            border: Border.all(
+                              color: const Color(0xFF6c6c6c),
+                              width: 0.15,
                             ),
-                            margin: const EdgeInsets.symmetric(vertical: 1.0),
-                            padding: appBarPadding,
-                            child: _OpenNavigationPane(
-                              theme: theme,
-                              pane: pane,
-                              paneKey: _panelKey,
-                              listKey: _listKey,
-                              onToggle: toggleCompactOpenMode,
-                              onItemSelected: toggleCompactOpenMode,
-                            ),
+                            borderRadius: BorderRadius.circular(8.0),
+                          ),
+                          margin: const EdgeInsets.symmetric(vertical: 1.0),
+                          padding: appBarPadding,
+                          child: _OpenNavigationPane(
+                            theme: theme,
+                            pane: pane,
+                            paneKey: _panelKey,
+                            listKey: _listKey,
+                            onToggle: toggleCompactOpenMode,
+                            onItemSelected: toggleCompactOpenMode,
                           ),
                         ),
                       );
@@ -456,8 +466,8 @@ class NavigationViewState extends State<NavigationView> {
                   key: _overlayKey,
                   duration: theme.animationDuration ?? Duration.zero,
                   curve: theme.animationCurve ?? Curves.linear,
-                  start: _minimalPaneOpen ? 0.0 : -_kOpenNavigationPanelWidth,
-                  width: _kOpenNavigationPanelWidth,
+                  start: _minimalPaneOpen ? 0.0 : -kOpenNavigationPaneWidth,
+                  width: kOpenNavigationPaneWidth,
                   height: MediaQuery.of(context).size.height,
                   child: PaneScrollConfiguration(
                     child: ColoredBox(
@@ -523,7 +533,7 @@ class NavigationViewState extends State<NavigationView> {
     return PrimaryScrollController(
       controller: scrollController,
       child: ScrollConfiguration(
-        behavior: const _NavigationViewScrollBehavior(),
+        behavior: const NavigationViewScrollBehavior(),
         child: child,
       ),
     );
@@ -641,7 +651,7 @@ class NavigationAppBar with Diagnosticable {
       } else {
         return const SizedBox.shrink();
       }
-      widget = SizedBox(width: _kCompactNavigationPanelWidth, child: widget);
+      widget = SizedBox(width: kCompactNavigationPaneWidth, child: widget);
       return widget;
     });
   }
@@ -659,6 +669,9 @@ class _NavigationAppBar extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    assert(debugCheckHasFluentLocalizations(context));
+    assert(debugCheckHasDirectionality(context));
+
     final direction = Directionality.of(context);
     final PaneDisplayMode displayMode =
         InheritedNavigationView.maybeOf(context)?.displayMode ??
@@ -671,10 +684,7 @@ class _NavigationAppBar extends StatelessWidget {
         return AnimatedPadding(
           duration: theme.animationDuration ?? Duration.zero,
           curve: theme.animationCurve ?? Curves.linear,
-          padding: [PaneDisplayMode.minimal, PaneDisplayMode.open]
-                  .contains(displayMode)
-              ? EdgeInsets.zero
-              : const EdgeInsets.only(left: 24.0),
+          padding: const EdgeInsetsDirectional.only(start: 16.0),
           child: DefaultTextStyle(
             style:
                 FluentTheme.of(context).typography.caption ?? const TextStyle(),
@@ -701,24 +711,16 @@ class _NavigationAppBar extends StatelessWidget {
       case PaneDisplayMode.minimal:
       case PaneDisplayMode.open:
       case PaneDisplayMode.compact:
-        final isMinimalPaneOpen =
-            InheritedNavigationView.maybeOf(context)?.minimalPaneOpen ?? false;
-        final double width =
-            displayMode == PaneDisplayMode.minimal && !isMinimalPaneOpen
-                ? 0.0
-                : displayMode == PaneDisplayMode.compact
-                    ? _kCompactNavigationPanelWidth
-                    : _kOpenNavigationPanelWidth;
         result = Stack(children: [
-          Row(children: [
+          Row(mainAxisSize: MainAxisSize.min, children: [
             leading,
             if (additionalLeading != null) additionalLeading!,
-            Expanded(child: title),
+            Flexible(child: title),
           ]),
           if (appBar.actions != null)
             Positioned.directional(
               textDirection: direction,
-              start: width,
+              start: 0,
               end: 0.0,
               top: 0.0,
               bottom: 0.0,
@@ -740,8 +742,12 @@ class _NavigationAppBar extends StatelessWidget {
   }
 }
 
-class _NavigationViewScrollBehavior extends FluentScrollBehavior {
-  const _NavigationViewScrollBehavior();
+/// The [ScrollBehavior] used on [NavigationView]
+///
+/// It generates a [Scrollbar] using the global scroll controller provided by
+/// [NavigationView]
+class NavigationViewScrollBehavior extends FluentScrollBehavior {
+  const NavigationViewScrollBehavior();
 
   @override
   Widget buildScrollbar(context, child, details) {